refactor: implement ShareGuestSignUpView component with guest signup functionality
