+++ src/views/Signup/index.ts
@@ -1,4 +1,2 @@
 export { default } from './SignupView';
-export * from './components';
-export * from './hooks';
-export * from './utils';
+export { default as ShareGuestSignUpView } from './ShareGuestSignUpView';
